Moved GremlinScriptChecker to gremlin-core
This class no longer relies on Groovy classes and has general applicablity to processing gremlin-language so it was better off in gremlin-core. CTR
